Makefile.am: fix missing files on "clean" target.
As seen on [Debian bug #1049666], running the "clean" target under
certain circumstances may cause the rule to fail with:

	rm: cannot remove 'test-reads.fq.gz': No such file or directory

This is because the rm commands are not forcefully run, with flag -f.
One way to address this is to use the built-in $(RM) from default make
variables, which typically will result in running rm -f instead of
plain rm, in turn resulting in no failure if the file hasn't been
generated in the first place.

If this variable were to not to be trusted for whatever reason, there
is also the option to set it explicitly first, e.g. `RM = rm -vf`.

[Debian bug #1049666]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1049666

Signed-off-by: Étienne Mollier <emollier@debian.org>